requirements.txt
tqdm
gitpython
